Split agents module
